% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FitGene.R
\docType{methods}
\name{coef,FitGene-method}
\alias{coef,FitGene-method}
\title{SPLS coefficient estimates for genes}
\usage{
\S4method{coef}{FitGene}(object)
}
\arguments{
\item{object}{output of selectGene function}
}
\description{
SPLS coefficient estimates for genes.
}
