% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FitPath.R
\docType{methods}
\name{coef,FitPath-method}
\alias{coef,FitPath-method}
\title{LASSO coefficient estimates for pathways}
\usage{
\S4method{coef}{FitPath}(object)
}
\arguments{
\item{object}{output of selectPath function}
}
\description{
LASSO coefficient estimates for pathways.
}
