% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FitGene-class}
\alias{FitGene-class}
\title{An S4 class to represent gene-level analysis results}
\description{
An S4 class to represent gene-level analysis results
}
\section{Slots}{

\describe{
\item{\code{score}}{list of SPLS latent components for each pathway}

\item{\code{geneSelected}}{list of selected genes in each pathway}

\item{\code{fit}}{list of SPLS fits}

\item{\code{dataPrefiltered}}{list of prefiltered data for each pathway}

\item{\code{inputdata}}{input data for prefilter function}
}}

