% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.R
\name{prefilter}
\alias{prefilter}
\title{prefilter function}
\usage{
prefilter(data, time, status, p.cut = 0.5, plist = plist)
}
\arguments{
\item{data}{A N by P dataframe, where N is number of observations, P is number of genes}

\item{time}{time passed to selectGene's time argument.}

\item{status}{status selectGene's status argument.}

\item{p.cut}{p.cut is the pvalue cutoff point for prefiltering}

\item{plist}{user provided gene lists organized in pathways}
}
\description{
a supervised prefiltering by fitting a cox regression model on the expression measures of each mRNA.  Only genes with p values smaller than a pre-selected cut-off point were included in further analysis. Here we chose p=0.5 as default cut-off point.
}
\examples{
data(TCGA)
prefilter.results=prefilter( data=TCGA$geneexpr, time=TCGA$t, status=TCGA$d, plist=TCGA$pathList )
}
