% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Prefiltered-class}
\alias{Prefiltered-class}
\title{An S4 class to represent pre-filtered data.}
\description{
An S4 class to represent pre-filtered data.
}
\section{Slots}{

\describe{
\item{\code{xlist}}{list of genes in each pathway}

\item{\code{inputdata}}{input data for prefilter function}
}}

