% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGene.R
\docType{methods}
\name{selectGene,Prefiltered-method}
\alias{selectGene,Prefiltered-method}
\alias{selectGene}
\title{Gene selection using SPLS model}
\usage{
\S4method{selectGene}{Prefiltered}(object, fold = 5, K = 5,
  etas = seq(0.1, 0.9, 0.1), seed = 123)
}
\arguments{
\item{object}{output list of prefilter step}

\item{fold}{The number of folds to use to perform the cross-validation process.}

\item{K}{the maximum number of hidden features in spls.}

\item{etas}{Thresholding parameter. eta should be between 0 and 1.}

\item{seed}{random seed that was set,  default = 123.}
}
\description{
Sparse Partial Least Square (SPLS) step for gene selection and dimension reduction.By applying SPLS to each pathway, we achieve the goal of gene selection and dimension reduction at the same time.
}
\examples{
data(TCGA)
prefilter.results=prefilter( data=TCGA$geneexpr, time=TCGA$t, status=TCGA$d, plist=TCGA$pathList )
gene.results=selectGene( object=prefilter.results, fold=5, K=5, etas=c(0.1,0.9),seed=123)
}
