% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FitGene.R
\docType{methods}
\name{show,FitGene-method}
\alias{show,FitGene-method}
\title{Summary of selectGene results}
\usage{
\S4method{show}{FitGene}(object)
}
\arguments{
\item{object}{output of selectGene function}
}
\description{
Summary of selectGene results.
}
