% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FitPath.R
\docType{methods}
\name{show,FitPath-method}
\alias{show,FitPath-method}
\title{Summary of selectPath results}
\usage{
\S4method{show}{FitPath}(object)
}
\arguments{
\item{object}{output of selectPath function}
}
\description{
Summary of selectPath results.
}
