% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Prefiltered.R
\docType{methods}
\name{show,Prefiltered-method}
\alias{show,Prefiltered-method}
\title{Prefiltering summary}
\usage{
\S4method{show}{Prefiltered}(object)
}
\arguments{
\item{object}{output of prefilter function}
}
\description{
Summary of prefiltering results.
}
