% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\docType{data}
\name{TCGA}
\alias{TCGA}
\title{TCGA data}
\format{A \code{list} object with four elements, \code{geneexpr} (z scores for gene expression data), \code{t} (survival time), \code{d} (survival status indicator), \code{pathList} (pathway information).}
\usage{
TCGA
}
\description{
example data
}
\examples{
data(TCGA)
}
\references{
TCGA
}
\keyword{datasets}
